--- src/components/CardGrid/CardGrid.tsx
@@ -1,9 +1,0 @@
-import React from "react"
-
-import * as styles from "./CardGrid.module.css"
-
-const CardGrid: React.FC<React.PropsWithChildren> = ({ children }) => {
-  return <div className={styles.cardGrid}>{children}</div>
-}
-
-export default CardGrid
