--- src/components/CardGrid/index.ts
@@ -1,3 +1,0 @@
-import CardGrid from "./CardGrid"
-
-export default CardGrid
