+++ src/components/ContentGrid/index.ts
@@ -1,0 +1,3 @@
+import ContentGrid from "./ContentGrid"
+
+export default ContentGrid
