+++ src/components/ContentLink/index.ts
@@ -1,0 +1,3 @@
+import ContentLink from "./ContentLink"
+
+export default ContentLink
