--- src/components/GitHubCard/index.ts
@@ -1,3 +1,0 @@
-import GitHubCard from "./GitHubCard"
-
-export default GitHubCard
