+++ src/components/GitHubShowcase/index.ts
@@ -1,0 +1,3 @@
+import GitHubShowcase from "./GitHubShowcase"
+
+export default GitHubShowcase
