--- src/components/PostCard/index.ts
@@ -1,3 +1,0 @@
-import PostCard from "./PostCard"
-
-export default PostCard
