+++ src/pages/blog/index.tsx
@@ -3,7 +3,7 @@ import type { HeadFC, PageProps } from "gatsby"
 import { graphql } from "gatsby"
 
 import Layout from "../../components/Layout"
-import PostCard from "../../components/PostCard"
+import ContentLink from "../../components/ContentLink"
 import SEO from "../../components/SEO"
 import Section from "../../components/Section"
 
@@ -12,7 +12,7 @@ const BlogPage: React.FC<PageProps<Queries.BlogPostsQuery>> = ({ data }) => {
     <Layout>
       <Section title="Posts" first>
         {data.allMdx.nodes.map(node => (
-          <PostCard
+          <ContentLink
             id={node.id}
             title={node.frontmatter?.title!}
             description={node.excerpt}
